refactor: reorganize chat-related types and imports for clarity
